refactor: fix import system
This commit removes all `sys.path.append` occurences in the simulation
module. It does keep those calls for unittests. `vector.py`,
`channel.py` and parts of `waveform.py` in `sdr_utils` are duplicated into
`simulation.modem.util`. This was necessary to fix imports.

All imports are explicit relative imports now.

PEP8 autoformat was applied in a lot of places.

The README is updated to reflect the actual module usage.